separate device utils in daal4py from common utils
