+++ daal4py/sklearn/decomposition/_pca.py
@@ -25,7 +25,7 @@ from sklearn.utils.extmath import stable_cumsum
 
 import daal4py
 from .._utils import getFPType, get_patch_message, sklearn_check_version
-from .._utils import support_usm_ndarray
+from .._device_offload import support_usm_ndarray
 import logging
 
 if sklearn_check_version('0.22'):
