+++ daal4py/sklearn/linear_model/logistic_path.py
@@ -49,8 +49,8 @@ from sklearn.linear_model._logistic import (
 from sklearn.preprocessing import LabelEncoder, LabelBinarizer
 from .._utils import (getFPType,
                       get_patch_message,
-                      sklearn_check_version,
-                      support_usm_ndarray)
+                      sklearn_check_version)
+from .._device_offload import support_usm_ndarray
 import logging
