+++ daal4py/sklearn/metrics/_ranking.py
@@ -26,7 +26,7 @@ from sklearn.preprocessing import label_binarize
 
 from ..utils.validation import _daal_assert_all_finite
 from .._utils import get_patch_message, sklearn_check_version
-from .._utils import support_usm_ndarray
+from .._device_offload import support_usm_ndarray
 import logging
 
 if sklearn_check_version('0.22'):
